Reduce amount of warnings to zero
The compilation of example will still
produce some warnings, but they are
unrelated to the new code
